[python][run] fixes bugs in new management of data
run fixes bug

+++ src/tools.py
@@ -10,9 +10,6 @@ from datetime import datetime
 import torch
 import shutil
 
-info_file = "../.info"
-best_file = "../.best"
-
 def rand_uint32():
     return np.random.randint(np.iinfo(np.int32).max, dtype=np.uint32).item()
 
@@ -58,6 +55,8 @@ def make_init_moves(mcts0, mcts1):
     return
 
 def get_unique(data):
+    """ averages duplicate board positions
+    """
     board, idc, inv, cnt = np.unique(data['board'], axis=0,
             return_index=True,
             return_inverse=True,
@@ -77,6 +76,8 @@ def get_unique(data):
     return np.stack(result)
 
 def get_rotations(data):
+    """ creates rotations of board
+    """
     rotations = []
     dt = data.dtype
     shape = data['board'].shape[-1]
@@ -97,39 +98,58 @@ def get_rotations(data):
 
     return np.concatenate([data, np.stack(rotations)])
 
-def info_train():
+def get_new_params():
+    """ returns file name where new parameters will be saved
+    """
+    params = "{}/{}{}_{}.pyt".format(
+                PARAM_PATH, MODEL_CLASS, SHAPE, datetime.now().strftime("%m%d_%H-%M-%S"))
+    return params
+
+def get_params():
+    """ returns parameters to best model
+    """
     if 'PARAM' in os.environ:
-        best = get_param_file(os.environ['PARAM'])
+        params = get_param_file(os.environ['PARAM'])
     else:
-        best = get_best()
+        params = get_best()
+        if params is None:
+            print("No parameters found")
+    return params
 
-    if best is None:
-        best = get_latest()
-        if best is None:
-            raise RuntimeError("No model to train")
-        set_best(best)
+def get_versus():
+    """ returns parameters against which will be best model playing
+    """
+    if 'VERSUS' in os.environ:
+        params = get_param_file(os.environ['VERSUS'])
+    else:
+        params = get_latest()
+        if params is None:
+            print("No parameters found")
+    return params
 
-    new_param_file = "{}/{}{}_{}.pyt".format(
-            PARAM_PATH, MODEL_CLASS, SHAPE, datetime.now().strftime("%m%d_%H-%M-%S"))
+def get_data():
+    """ returns list of filenames containing data for training
+        also removes obsolete data
+    """
     if 'DATA' in os.environ:
         data = os.environ['DATA'].split(',').strip()
     else:
-        data_dirs = glob("{}/{}{}_*".format(DATA_PATH, MODEL_CLASS, SHAPE))
-        data_dirs.sort()
-        if len(data_dirs) > 0:
-            generation = int(data_dirs[-1][-3:])
+        data = glob("{}/{}{}_*".format(DATA_PATH, MODEL_CLASS, SHAPE))
+        data.sort()
+        if len(data) > 0:
+            generation = int(data[-1][-3:])
         else:
             raise RuntimeError("No data for training found")
 
-        if len(data_dirs) > 4:
+        if len(data) > 4:
             window = min(WINDOW[0]+(generation-3)//WINDOW[2], WINDOW[1])
             print("Window size is {}".format(window))
-            print("Number of data dirs is {}".format(len(data_dirs)))
-            if len(data_dirs) - window > 0:
-                for d in data_dirs[0:-window]:
+            print("Number of data dirs is {}".format(len(data)))
+            if len(data) - window > 0:
+                for d in data[0:-window]:
                     logging.info("Removing file {}".format(f))
                     os.remove(f)
-        data = data_dirs[-window:]
+            data = data[-window:]
 
     data_files = []
     for d in data:
@@ -142,9 +162,11 @@ def info_train():
         else:
             data_files.append(os.path.realpath(d))
 
-    return best, new_param_file, data_files
+    return data_files
 
-def new_generation():
+def init_generation():
+    """ creats a directory for new generation of data
+    """
     data_dirs = glob("{}/{}{}_*/".format(DATA_PATH, MODEL_CLASS, SHAPE))
     data_dirs.sort()
     if len(data_dirs) == 0:
@@ -173,53 +195,40 @@ def new_generation():
 
     return data_dir
 
-def info_selfplay():
-    """ returns best params and new data file name
-        must return valid values!!
+def get_new_data():
+    """ returns name of new data file
     """
-    if 'PARAM' in os.environ:
-        best = get_param_file(os.environ['PARAM'])
-    else:
-        best = get_best()
-
+    # if self-play is runnig in multiple processes
+    # setting sequence number will avoid name colisions
     if 'SEQUENCE' in os.environ:
         seq = os.environ['SEQUENCE']
     else:
         seq = 0
 
-    if best is None:
-        best = get_latest()
-    if best is None:
-        best = "{}/{}{}_{}.pyt".format(
-                PARAM_PATH, MODEL_CLASS, SHAPE, datetime.now().strftime("%m%d_%H-%M-%S"))
-
+    # finding directory where will be data saved
     data_dir = glob("{}/{}{}_*/".format(DATA_PATH, MODEL_CLASS, SHAPE))
     data_dir.sort()
     if len(data_dir) == 0:
-        data_dir = new_generation()
+        data_dir = init_generation()
     else:
         data_dir = data_dir[-1]
 
     file_name = "{}/{}s{}".format(data_dir, datetime.now().strftime("%m%d_%H-%M-%S"), seq)
 
-    return best, file_name
+    return file_name
 
-def info_eval():
-    if 'PARAM' in os.environ != 'VERSUS' in os.environ:
-        raise RuntimeError("Only one out of -p|--param and -v|--versus specified. Specify both or none")
-    if 'PARAM' in os.environ and 'VERSUS' in os.environ:
-        print( get_param_file(os.environ['PARAM1']), get_param_file(os.environ['PARAM1']), True)
-        sys.exit()
-        return get_param_file(os.environ['PARAM1']), get_param_file(os.environ['PARAM1']), True
-    best = get_best()
-    latest = get_latest()
-    if best == latest:
-        return None, latest
-    print(best, latest, False)
-    sys.exit()
-    return best, latest, False
+def dry_run():
+    """ if dry run is set best parameters won't be changed
+        returns True if dry run is set
+                False if not
+    """
+    if 'DRY_RUN' in os.environ:
+        return True
+    return False
 
 def get_latest():
+    """ returns latest parameters
+    """
     param_files = glob("{}/{}{}_*.pyt".format(
         PARAM_PATH, MODEL_CLASS, SHAPE))
     if len(param_files) == 0:
@@ -230,6 +239,9 @@ def get_latest():
     return param_files[-1]
 
 def get_param_file(name):
+    """ function will check if parameters exists
+        returns absolute path to parameters
+    """
     if os.path.isfile(name):
         return name
     abs_name = "{}/{}".format(PARAM_PATH, os.path.basename(name))
@@ -239,6 +251,8 @@ def get_param_file(name):
         raise RuntimeError("File {} or {} does not exists".format(name, abs_name))
 
 def get_best():
+    """ returns best parameters
+    """
     with open(PARAM_BEST, "r+") as fp:
         try:
             content = json.load(fp)
@@ -253,12 +267,18 @@ def get_best():
             path = os.path.basename(content['path'])
             path = os.path.join(PARAM_PATH, path)
             if not os.path.isfile(path):
-                raise RuntimeError("File {} does not exst".format(content['path']))
+                print("Could not found parameters")
+                print("Running with default parameters")
             return path
 
         return content['path']
 
 def set_best(path):
+    """ sets new best parameters
+    """
+    if dry_run():
+        print("Dry run: setting {} as best".format(path))
+        return
     with open(PARAM_BEST, "w") as fp:
         if not os.path.isabs(path):
             path = "{}/{}".format(PARAM_PATH, path)
@@ -270,7 +290,9 @@ def set_best(path):
         json.dump(content, fp)
 
 def init_param_file():
-    open(PARAM_BEST, 'a').close
+    """ initialises file containing path to file with best parameters
+    """
+    open(PARAM_BEST, 'a').close()
 
 def self_play_config():
     s = ""
